Wait for slots instead of TaskManagers in test environment
Instead of waiting for the number of connected task managers, the local
mini cluster used for tests now waits for the total number of available
slots. Waiting for the number of connected task managers instead of the
available slots might result in races in rare situations.

In addition, rename task tracker (sic) to task manager in test classes.