[FLINK-1463] Fix stateful/stateless Serializers and Comparators
Before, Serializers would announce whether they are stateful or not and
rely on RuntimeStatefulSerializerFactory to do the duplication.
Comparators, on the other hand, had a duplicate method that the user was
required to call.

This commit removes the statful/stateless property from Serializers but
instead introduces a duplicate() method, similar to Comparators, that
can return the same instance.

The two serializer factories are merged into one that always calls
duplicate() before returning a serializer.