[runtime] Fix scheduleOrUpdateConsumers logic for blocking results
For tasks producing mixed pipelined and blocking results, the scheduling of
receivers could lead to deadlocks, because of missing notifications.

This problem was discovered in the ALSITCase, where the iteration head is
producing pipelined results for the step function and blocking results for
the final results. The receivers of the final blocking result never got
notified.

This commit adds a separate test case to test the correct behaviour and reverts
the changes introduced in dafcd4e to force pipelining in ALSITCase.