[runtime] Fix Netty watermark configuration
The watermarks of the NettyServer, which is responsible for data transfers, were
not configured correctly. Instead they were configured for the NettyClient, which
does need this fine-grained control as it does not transfer a lot of data.

Previously, the default watermarks matched our desired watermarks of 64k (for our
default page size) by chance.