[FLINK-2852] [test-stability] Fix ScalaShellITSuite and ScalaShellLocalStartupITCase
Changes test program to use an int accumulator which is checked at the end of the program.
This avoids to look for the "Job status changed to FINISHED." string in the stdout output,
which can sometimes not be printed if the JobExecutionResult arrives earlier than the
JobStatusChanged message at the JobClientActor.