[hotfix] Fix shaky EventTimeAllWindowCheckpointITCase
In very rare cases it could happen that a checkpoint would be performed
after the ValidatingSink signaled that it had seen all expected
elements. If this happened the job would be restarted with the already
complete state and we would never finish since no more elements would
arrive.

This adds a check in open() of ValidatingSink that signals success if we
already have the final state.