[FLINK-2964] [runtime] Fix broken spilling of MutableHashTable
The HashPartition did not count properly the number of occupied memory
segments, because it excluded the memory segments of the
BuildSideBuffer. That caused the MutableHashTable to fail when trying to
spill a partition which did not have any overflow segments. This PR
fixes the problem by also counting the memory segments of the
BuildSideBuffer.

This closes #1324.