Revert "[FLINK-3369] [runtime] Make RemoteTransportException instance of CancelTaskException"
This reverts commit cf3ae88b73e30a2d69ac1cc6009a8304ea3f53cc.

The reason is that we cannot be sure that a failing sender will always be the root
cause why the receiver cannot read the data. A TM can also fail after a sender is
finished leading to a loss of the intermediate result partitions. Then the receiver
will no longer be able to read the data.