[FLINK-3824] ResourceManager may repeatedly connect to outdated JobManager
When the ResourceManager receives a new leading JobManager via the
LeaderRetrievalService it tries to register with this JobManager until
connected. If during registration a new leader gets elected, the
ResourceManager may still repeatedly try to register with the old
one. This doesn't affect the registration with the new JobManager but
leaves error messages in the log file and may process unnecessary
messages.